requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.7.0
bleach==6.0.0
brotlipy==0.7.0
build==0.10.0
bump2version==1.0.1
bumpversion==0.6.0
CacheControl==0.12.14
Cerberus==1.3.5
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
colorama==0.4.6
constantly==15.1.0
crashtest==0.4.1
cryptography==41.0.3
cssselect==1.2.0
decorator==5.1.1
distlib==0.3.7
Django==4.1.5
django-appconf==1.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.20.1
dulwich==0.21.5
extruct==0.16.0
filelock==3.12.2
fire==0.5.0
frozenlist==1.4.0
gunicorn==20.1.0
h11==0.14.0
html-text==0.5.2
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
iniconfig==2.0.0
installer==0.7.0
isodate==0.6.1
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
jaraco.classes==3.3.0
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jstyleson==0.0.2
keyring==23.9.0
lockfile==0.12.2
loguru==0.7.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mf2py==1.1.3
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.8
packaging==23.1
parsel==1.8.1
pathspec==0.11.2
pdoc3==0.10.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.10.0
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
Protego==0.3.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyproject_hooks==1.0.0
pyRdfa3==3.5.3
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
queuelib==1.6.2
rapidfuzz==2.15.1
rcssmin==1.1.1
rdflib==7.0.0
readme-renderer==40.0
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.5.2
rjsmin==1.2.1
rpds-py==0.9.2
ruff==0.0.269
scrapfly-sdk==0.8.8
Scrapy==2.10.0
SecretStorage==3.3.3
sentry-sdk==1.29.2
service-identity==23.1.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
termcolor==2.2.0
tldextract==3.4.4
tomlkit==0.12.1
tqdm==4.66.1
trove-classifiers==2023.8.7
twine==4.0.2
Twisted==22.10.0
typing_extensions==4.7.1
urllib3==1.26.11
virtualenv==20.24.3
w3lib==2.1.2
webencodings==0.5.1
whitenoise==6.3.0
yarl==1.9.2
zipp==3.16.2
zope.interface==6.0